'Fix' the goto scroll issue caused by the version upgrade.
The core of the issue seems to be that Egui tries to align a row to the top of the scroll area (as ordered), but runs out of elements to pad the bottom with, and thus kind of gives up. Note, this only happens if we're at the end of the address space (as then we no longer have new values to pad the bottom with).

This is a new issue in Egui 0.14, as previous versions made a best effort attempt to still scroll to the desired row, even if they couldn't align it to the top of the screen.

By now instead aligning to the Center of the screen it once again seems to do a best effort scroll, and scroll to the row we want, even if it can't quite align it at the centre of the screen.